Added separate interface ContextAware which contains the isTestContext function
