+++ src/main/kotlin/io/github/manamiproject/modb/core/Functions.kt
@@ -1,5 +1,6 @@
 package io.github.manamiproject.modb.core
 
+import io.github.manamiproject.modb.core.config.ContextAware
 import io.github.manamiproject.modb.core.config.MetaDataProviderConfig
 import io.github.manamiproject.modb.core.coroutines.ModbDispatchers.LIMITED_CPU
 import io.github.manamiproject.modb.core.coroutines.ModbDispatchers.LIMITED_FS
@@ -86,12 +87,12 @@ public fun random(number1: Int, number2: Int): Long {
 
 /**
  * Only executes the given function if the current context is not the test context.
- * @since 8.0.0
- * @param config Config of a meta data provider.
+ * @since 14.0.0
+ * @param contextAwareImpl Config of a meta data provider.
  * @param func Function to be executed if the current context is not the test context.
  */
-public suspend fun excludeFromTestContext(config: MetaDataProviderConfig, func: suspend () -> Unit): Unit = withContext(LIMITED_CPU) {
-    if (!config.isTestContext()) {
+public suspend fun excludeFromTestContext(contextAwareImpl: ContextAware, func: suspend () -> Unit): Unit = withContext(LIMITED_CPU) {
+    if (!contextAwareImpl.isTestContext()) {
         func.invoke()
     }
 }
